Tornasole hook from config json (#104)
* creating tornasole hook from config

* making a quick variance fix (#99)

* Adding the change to convert ndarray to np.ndarray when operator is not available in mxnet.

* Cleanup and tests for TF and mxnet

* remove rmtree from s3 test

* Fixed the function invocation of get_numpy_reduction

* Changes to read from hardcoded path

* fixing pytorch test

* Setting SaveConfig per mode (#94)

* add doc for passing saveconfig specific to modes

* add save config for collection

* Create an option to build tornasole with no framework, TORNASOLE_FOR_RULES=1 (#95)

* add option to build only for rules

* Adding support to set save config per mode through json, also copying load collection method to all frameworks as that was missed

* remove set -ex from tests script since it prevents upload of reports

* move json config out of hooks

+++ tests/tensorflow/hooks/test_save_all_full.py
@@ -1,19 +1,23 @@
 from .utils import *
 from tornasole.tensorflow import reset_collections, get_collections
-import pytest
 import shutil, glob
 from tornasole.core.reader import FileReader
+from tornasole.core.json_config import TORNASOLE_CONFIG_FILE_PATH_ENV_STR
 
-def test_save_all_full():
+
+def test_save_all_full(hook=None, trial_dir=None):
   run_id = 'trial_'+datetime.now().strftime('%Y%m%d-%H%M%S%f')
-  trial_dir = os.path.join(TORNASOLE_TF_HOOK_TESTS_DIR, run_id)
-
+  if trial_dir is None:
+    trial_dir = os.path.join(TORNASOLE_TF_HOOK_TESTS_DIR, run_id)
   tf.reset_default_graph()
-  reset_collections()
-
-  hook = TornasoleHook(out_dir=trial_dir,
+  hook_created = False
+  if hook is None:
+    reset_collections()
+    hook = TornasoleHook(out_dir=trial_dir,
                        save_all=True,
                        save_config=SaveConfig(save_interval=2))
+    hook_created = True
+
   simple_model(hook)
   _, files = get_dirs_files(trial_dir)
   dirs, _ = get_dirs_files(os.path.join(trial_dir, 'events'))
@@ -54,4 +58,14 @@ def test_save_all_full():
         size += tensor_data.nbytes
     assert i == 85
     assert size == 1470
-  shutil.rmtree(trial_dir)
+  if hook_created:
+    shutil.rmtree(trial_dir)
+
+def test_hook_config_json():
+  out_dir = 'newlogsRunTest1/test_hook_from_json_config'
+  shutil.rmtree(out_dir, ignore_errors=True)
+  os.environ[TORNASOLE_CONFIG_FILE_PATH_ENV_STR] = 'tests/tensorflow/hooks/test_json_configs/test_hook_from_json_config.json'
+  reset_collections()
+  hook = TornasoleHook.hook_from_config()
+  test_save_all_full(hook, out_dir)
+  shutil.rmtree(out_dir, ignore_errors=True)
